pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["kubrain", "kubrain.*"]
namespaces = true

[project]
name = "kubrain"
version = "0.1.0"
description = "A framework for deploying local ML models on private K8s clusters"
license = {file = "LICENSE"}
dependencies = [
    "PyYAML==6.0.1"
]
